Use Cases: Improve files architecture
