+++ src/use_cases/helpers/mod.rs
@@ -1,2 +1,4 @@
+pub(crate) mod csv;
 pub(crate) mod language;
 pub(crate) mod localized_item;
+pub(crate) mod model;
