+++ src/use_cases/import_delivery_slip.rs
@@ -7,9 +7,16 @@ use crate::{
             models::delivery_slip::{batch_upsert, DeliverySlipModel},
         },
     },
+    interface_adapters::mappers::CSVToEntityParser,
 };
 
-use super::*;
+use super::{
+    helpers::{
+        csv::{CanReadCsvUseCase, ImportCsvUseCase},
+        model::CanPersistIntoDatabaseUseCase,
+    },
+    *,
+};
 
 #[derive(Default)]
 pub struct ImportDeliverySlipUseCase {
